requirements-dev.txt
# install all the project requirements
-r requirements.txt

# intall the requirements for test
pytest==6.2.5
pytest-cov==3.0.0
mock==2.0.0
fakeredis==1.7.4
sqlalchemy-utils
